Make Config constructor to call the ActorSystemSetup version
